% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEAUTO.R
\docType{data}
\name{BASEAUTO}
\alias{BASEAUTO}
\title{Claims severity from car insurance domain}
\format{
\subsection{\code{BASEAUTO}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
BASEAUTO
}
\description{
A subset of data from a French company
Report ...
}
\keyword{datasets}
