% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodedBASEAUTO_new.R
\docType{data}
\name{encodedBASEAUTO_new}
\alias{encodedBASEAUTO_new}
\title{Claims severity from car insurance domain (encoded from BASEAUTO)}
\format{
\subsection{\code{encodedBASEAUTO_new}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
encodedBASEAUTO_new
}
\description{
A subset of data from a French company
Report ...
}
\keyword{datasets}
